Make Array#map, Array#some, Array#every and Array#filter implementations (almost) ES5 compliant and use native versions when possible, yielding huge performance improvements.
Prototype's implementations of these methods and the ES5 spec differ in the way a missing iterator function is handled:
Prototype uses Prototype.K as a default iterator, while ES5 specifies that a TypeError has to be thrown. Implementing the ES5 spec completely would break backward compatibility of Prototype and would force users to pass Prototype.K manually.

To keep backward compatibility when using browser native methods for Array#every and so on, these methods get wrapped. Some early performance benchmarks show that this wrapping has almost no performance impact.

This change should also fix #765, as the fallback implementations for the ES5 methods all can be called in a generic way.