Change failing tests.
One of these tests presumably tests comma handling — whether we do the right thing when the space is before the comma, or when there end up being duplicate commas — and this is not something we handled before, so if this is a newly-failing test it's because Sizzle used to handle this and now does not. That's fine with me.

The other one appears to be a disagreement over whether `:enabled` should match `input[type=hidden]`, and Sizzle apparently says it does, and I'm OK with that, too.