Add a `test:server` rake task.
For IE, I find it far easier to start the test server on my dev machine, then use my IE machine to visit the test pages manually. For that I need a task that runs the server indefinitely.

At some point I'll add a page that does nothing but list the tests for convenience.