A first pass on a complete rewrite of the unit tests.
Not confident enough to get rid of the existing stuff, so the new folder is `test.new` and the new Rake tasks are under the `test_new` namespace.

Use `rake test_new:start` to start the test server, and then visit the main URL to run all tests. Or add a comma-separated list of test suites (e.g., http://127.0.0.1:4567/test/number,string) to run only certain tests.

There is a simple test runner, too, though I haven’t yet tested it on Windows or Linux. Run `rake test_new:run` to invoke the runner, using the same `BROWSERS` and `TESTS` environment variables that are used with the old `rake test`.